% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refresh.R
\name{nygrid_dqc}
\alias{nygrid_dqc}
\title{Data quality checks}
\usage{
nygrid_dqc(input)
}
\arguments{
\item{input}{A nygrid object (output of the nygrid_dr function)}
}
\value{
A data.frame object with the tests results
}
\description{
The function conducts a set of data quality tests for the NYgrid dataset
}
