% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refresh.R
\name{nygrid_dr}
\alias{nygrid_dr}
\title{Refresh the NYgrid Dataset}
\usage{
nygrid_dr(start, end, offset, parent, subba, api_key, api_path)
}
\arguments{
\item{start}{The query start time in POSIXt format}

\item{end}{The query end time in POSIXt format}

\item{parent}{The series parent name}

\item{subba}{The sub-region codes}

\item{api_key}{A string, EIA API key, see https://www.eia.gov/opendata/ for registration to the API service}

\item{api_path}{A string, the API path to follow the API endpoint https://api.eia.gov/v2/.
The path can be found on the EIA API dashboard, for more details see https://www.eia.gov/opendata/browser/}
}
\value{
A data.frame object
}
\description{
The function pull data the NYgrid dataset from the EIA API based
on time range using the EIAapi eia_backfill function
}
