% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nyis_subba_codes}
\alias{nyis_subba_codes}
\title{New York Sub Regions Codes}
\usage{
nyis_subba_codes()
}
\value{
A list of sub-region codes
}
\description{
A list of New York Independent System Operator sub-regions codes
}
